done singup controller to 99%, updating styles for signup view

+++ client/components/panel/exploration/location/location.service.js
@@ -5,7 +5,8 @@
 	'use strict';
 
 	function LocationService($q, $http, Auth){
-		var deferred = null;
+		var deferred = null,
+		access_token = Auth.getToken();
 		return {
 			apiBaseURL: 'http://52.8.211.37/api.walmex.latlong.mx',
 
@@ -13,26 +14,27 @@
 			 * [addNewLocation Add new locations to map]
 			 * @param {[type]} formData [Serialized data from form]
 			 */
-			addNewLocation: function(formData){
-				var access_token = Auth.getToken();
-				deferred = $q.defer();
+			addNewLocation: function(formData){ 
+				if (access_token) {
+					deferred = $q.defer();
 
-				$http({
-					url: this.apiBaseURL+'/ws/places',
-					method: "POST",
-					data: formData,
-					transformRequest: angular.identity,
-					headers: {
-						'Content-Type': undefined,
-						'Authorization': 'Bearer '+access_token.access_token
-					}
-				})
-				.then(function(dat){
-					deferred.resolve(dat);
-				}, function(err){
-					deferred.reject(err);
-				});
-				return deferred.promise;
+					$http({
+						url: this.apiBaseURL+'/ws/places',
+						method: "POST",
+						data: formData,
+						transformRequest: angular.identity,
+						headers: {
+							'Content-Type': undefined,
+							'Authorization': 'Bearer '+access_token.access_token
+						}
+					})
+					.then(function(dat){
+						deferred.resolve(dat);
+					}, function(err){
+						deferred.reject(err);
+					});
+					return deferred.promise;
+				}
 			},
 
 			/**
@@ -41,25 +43,26 @@
 			 */
 			getLocations: function(opts){
 				deferred = $q.defer();
-				var access_token = Auth.getToken();
-				var _locations = $http({
-					url: this.apiBaseURL+'/ws/places',
-					method: "GET",
-					headers: {
-						'Content-Type': undefined,
-						'Authorization': 'Bearer '+access_token.access_token
-					},
-					params: opts?opts:''
-        });
-				_locations.then(function(result){
-					deferred.resolve(result);
-				}, function(error){
-					if(error.status===401 && error.statusText==='Unauthorized'){
-						//Actualizar token
-					}
-					deferred.reject(error);
-				});
-				return deferred.promise;
+				if (access_token) {
+					var _locations = $http({
+						url: this.apiBaseURL+'/ws/places',
+						method: "GET",
+						headers: {
+							'Content-Type': undefined,
+							'Authorization': 'Bearer '+access_token.access_token
+						},
+						params: opts?opts:''
+	        });
+					_locations.then(function(result){
+						deferred.resolve(result);
+					}, function(error){
+						if(error.status===401 && error.statusText==='Unauthorized'){
+							//Actualizar token
+						}
+						deferred.reject(error);
+					});
+					return deferred.promise;
+				}
 			},
 			getSingleLocation: function(id){
 				console.log(id)
@@ -89,49 +92,49 @@
 			 * @param  {[type]} id [Id of layer to remove]
 			 */
 			updateLocationLayer: function(id){
-				var access_token = Auth.getToken();
 				deferred = $q.defer();
-
-				var _locations = $http({
-					url: this.apiBaseURL+'/ws/places/' + id,
-					method: "GET",
-					headers: {
-						'Content-Type': undefined,
-						'Authorization': 'Bearer '+access_token.access_token
-					}
-        });
-				_locations.then(function(result){
-					deferred.resolve(result);
-				}, function(error){
-					if(error.status===401 && error.statusText==='Unauthorized'){
-						//Actualizar token
-					}
-					deferred.reject(error);
-				});
-				return deferred.promise;
+				if (access_token) {
+					var _locations = $http({
+						url: this.apiBaseURL+'/ws/places/' + id,
+						method: "GET",
+						headers: {
+							'Content-Type': undefined,
+							'Authorization': 'Bearer '+access_token.access_token
+						}
+	        });
+					_locations.then(function(result){
+						deferred.resolve(result);
+					}, function(error){
+						if(error.status===401 && error.statusText==='Unauthorized'){
+							//Actualizar token
+						}
+						deferred.reject(error);
+					});
+					return deferred.promise;
+				}
 			},
 
 			delLocation: function(id){
 				deferred = $q.defer();
-				var access_token = Auth.getToken();
-
-				var _locations = $http({
-					url: this.apiBaseURL+'/ws/places/'+id,
-					method: "DELETE",
-					headers: {
-						'Content-Type': undefined,
-						'Authorization': 'Bearer '+access_token.access_token
-					}
-				});
-				_locations.then(function(result){
-					deferred.resolve(result);
-				}, function(error){
-					if(error.status===401 && error.statusText==='Unauthorized'){
-						//Actualizar token
-					}
-					deferred.reject(error);
-				});
-				return deferred.promise;
+				if (access_token) {
+					var _locations = $http({
+						url: this.apiBaseURL+'/ws/places/'+id,
+						method: "DELETE",
+						headers: {
+							'Content-Type': undefined,
+							'Authorization': 'Bearer '+access_token.access_token
+						}
+					});
+					_locations.then(function(result){
+						deferred.resolve(result);
+					}, function(error){
+						if(error.status===401 && error.statusText==='Unauthorized'){
+							//Actualizar token
+						}
+						deferred.reject(error);
+					});
+					return deferred.promise;
+				}
 			},
 
 			/**
@@ -140,25 +143,25 @@
 			 * @param  {[type]} id_layer [description]
 			 */
 			updateLocationVar: function(formData, id_layer){
-				var access_token = Auth.getToken();
 				deferred = $q.defer();
-
-				$http({
-					url: this.apiBaseURL+'/ws/places_u/'+id_layer,
-					method: "POST",
-					data: formData,
-					transformRequest: angular.identity,
-					headers: {
-						'Content-Type': undefined,
-						'Authorization': 'Bearer '+access_token.access_token
-					}
-				})
-				.then(function(dat){
-					deferred.resolve(dat);
-				}, function(err){
-					deferred.reject(err);
-				});
-				return deferred.promise;
+				if (access_token) {
+					$http({
+						url: this.apiBaseURL+'/ws/places_u/'+id_layer,
+						method: "POST",
+						data: formData,
+						transformRequest: angular.identity,
+						headers: {
+							'Content-Type': undefined,
+							'Authorization': 'Bearer '+access_token.access_token
+						}
+					})
+					.then(function(dat){
+						deferred.resolve(dat);
+					}, function(err){
+						deferred.reject(err);
+					});
+					return deferred.promise;
+				}
 			}
 		};
 	}
